* tree-cfg.c (stmt_dominated_by_p): Handle statements defining initial 	value of a variable. 	* tree-ssa-loop-niter.c (stmt_dominates_stmt_p): Removed. 	(can_count_iv_in_wider_type_bound): Use stmt_dominated_by_p.
	* tree-pretty-print.c (dump_generic_node): Cast pointer to size_t
	first.
	* tree-ssa-operands.c (test_imm_print, verify_abort): Ditto.


git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/ssaupdate-branch@89508 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/tree-pretty-print.c
@@ -243,7 +243,7 @@ dump_generic_node (pretty_printer *buffer, tree node, int spc, int flags,
     dump_vops (buffer, node, spc, flags);
 
   if (is_stmt && (flags & TDF_STMT_ADDR))
-    pp_printf (buffer, "<&0x%x> ", (unsigned int)node);
+    pp_printf (buffer, "<&0x%x> ", (unsigned int)(size_t) node);
 
   if (dumping_stmts
       && (flags & TDF_LINENO)
